requirements.txt
Pillow>=10.0.0
piexif>=1.1.3
exifread>=3.0.0
folium>=0.14.0
pandas>=2.0.0
rich>=13.0.0
click>=8.1.0